Issue #6 - reorganize this package to make it cleaner and more extinsible. Isolate each function and make room to expand specific campaign logic. Separate the idea of campaign, variable, and profiles
