+++ insitupy/campaigns/snowex/__init__.py
@@ -1,0 +1,2 @@
+from .metadata_variables import SnowExMetadataVariables
+from .primary_variables import SnowExPrimaryVariables
