+++ insitupy/campaigns/snowex/snowex_campaign.py
@@ -1,0 +1,83 @@
+import pandas as pd
+from pathlib import Path
+
+from insitupy.io.metadata import MetaDataParser
+from . import SnowExMetadataVariables, SnowExPrimaryVariables
+from insitupy.variables import MeasurementDescription
+from insitupy.profiles.base import ProfileData, standardize_depth
+
+
+class SnowExMetadataParser(MetaDataParser):
+    METADATA_VARIABLE_CLASS = SnowExMetadataVariables
+    PRIMARY_VARIABLES_CLASS = SnowExPrimaryVariables
+
+
+class SnowExProfileData(ProfileData):
+    META_PARSER = SnowExMetadataParser
+
+    @classmethod
+    def from_file(cls, fname, variable: MeasurementDescription):
+        # TODO: timezone here (mapped from site?)
+        meta_parser = cls.META_PARSER(fname, "US/Mountain")
+        # Parse the metadata and column info
+        metadata, columns, header_pos = meta_parser.parse()
+        # read in the actual data
+        data = cls._read(fname, columns, header_pos)
+
+        return cls(data, metadata, variable)
+
+    @staticmethod
+    def _read(profile_filename, columns, header_position):
+        """
+        # TODO: better name mapping here
+        Read in a profile file. Managing the number of lines to skip and
+        adjusting column names
+
+        Args:
+            profile_filename: Filename containing the a manually measured
+                             profile
+        Returns:
+            df: pd.dataframe contain csv data with standardized column names
+        """
+        # header=0 because docs say to if using skip rows and columns
+        df = pd.read_csv(
+            profile_filename, header=0,
+            skiprows=header_position,
+            names=columns,
+            encoding='latin'
+        )
+        # Special SMP specific tasks
+        depth_fmt = 'snow_height'
+        is_smp = False
+        if 'force' in df.columns:
+            # Convert depth from mm to cm
+            df['depth'] = df['depth'].div(10)
+            is_smp = True
+            # Make the data negative from snow surface
+            depth_fmt = 'surface_datum'
+
+            # SMP serial number and original filename for provenance to the comment
+            f = Path(profile_filename).name
+            serial_no = f.split('SMP_')[-1][1:3]
+
+            df['comments'] = f"fname = {f}, " \
+                             f"serial no. = {serial_no}"
+
+        if not df.empty:
+            # Standardize all depth data
+            new_depth = standardize_depth(
+                df['depth'], desired_format=depth_fmt, is_smp=is_smp
+            )
+
+            if 'bottom_depth' in df.columns:
+                delta = df['depth'] - new_depth
+                df['bottom_depth'] = df['bottom_depth'] - delta
+
+            df['depth'] = new_depth
+
+            delta = abs(df['depth'].max() - df['depth'].min())
+            LOG.debug(
+                f'File contains a profile with'
+                f' with {len(df)} layers across {delta:0.2f} cm'
+            )
+        return df
