+++ insitupy/io/strings.py
